Fixed race condition by moving tunnels out of the link object, they are anyway independent.
This bug happened when node joining from two directions, one of the nodes discovers it and close one of the connections, and in some instances the tunnel objects were discarded at the same time. This then made a following clash between the discarded tunnel object and a simultaneous tunnel setup from other node go undetected.