Support for static actor deployment on node requirements added.
This is a first try to introduce support for requirements that guide the deployment of actors on nodes, so far only once directly after the application is launched. For example it is possible to specify that an actor needs to run on a node based on that node's attributes. Currently the deployment requirements are specified in a JSON encoded file.

cscontrol deploy got new option --reqs JSON-file
control API new POST /application/{application-id}/migrate interface
appmanager expanded with the deployment requirement logic, applying union, section and removal set operation between matching rules
appmanager also chose node based on the connections between actors
plugg-ins for matching rules that returns set of nodes