Correct Delay actor, add RecTimer
NOTE: std.Delay throws if asked to migrate

The Delay actor did not implement a delay. It was an implementation
of a recurrent timer but did not function as a tick either since it
didn't clear trigger on empty queue and didn't flush the input. A
recurrent timer actor has been added however in case that is what
someone intended. If the original operation of the Delay actor is
wanted it needs to be reimplemented under a different name.