Implements a storage proxy solution, where proxy client runtimes can configure a server storage proxy runtime.
For the proxy client runtime the configure parameter global.storage_proxy specifies a runtime to runtime uri, e.g. calvinip://<addr>:<port>, to a proxy server runtime.
This can be specified using environment variable CALVIN_GLOBAL_STORAGE_PROXY=\"calvinip://<addr>:<port>\"
All runtimes that are not proxy clients are proxy servers. The proxy client is implemented as a storage plugin.
The proxy server is implemented directly in storage higher level.
The communication is over a tunnel on the runtime to runtime link.

Example usage is first start a single runtime with default configuration, if it is the only runtime without proxy it will use its internal local storage.
Then start runtimes or run py.test while specifying the CALVIN_GLOBAL_STORAGE_PROXY variable referencing to the first runtime. This is good for verifying if the DHT storage is the cause of problems. Other usage is for smaller devices that can't handle to be part of the DHT.