Introduced proper HTTP status codes on the control API.
utils is updated to raise exception when status code not in success range (2xx).
Make a few more requests wait for a callback before responding if successful or not.
The ones not waiting only returns ACCEPTED and not OK.
The CalvinResponse class now have support for min and max, use max of a list of objects to find if any errors (will return with highest status code found)