Introduced shadow actors, used as a actor instance placeholder when the wanted actor class is not available or requirements are not fulfilled on the runtime it is deployed to. Added a storage command to cscontrol to read index values.
To allow shadow actors a new option is available --no-check in cscontrol deploy that allows actor classes that is not locally known.
A new ShadowActor class is defined.
Port manager creates missing ports on ShadowActor instances when not found.
Updated test_requirements.py with test of shadow actors.

Shadow components are NOT supported yet!