Added support for packet forwarding and usage of proxy runtimes
Added support for packet forwarding using forwarding-links, a
forwarding-link is requested by sending a ROUTE_REQUEST command
to a peer. The ROUTE_REQUEST command includes information about
the destination peer and if a connection exists or can be created
to the destination peer a forwarding-link is created and forwards
all communication to and from the destination and originator peer.

And added possibility to use a peer runtime as a proxy for all
runtime to runtime communication. A peer defines a proxy by setting
the node id of the proxy in a 'proxy' attribute in the node info
structure stored in the global registry.