Support for access control on control interface
The access controle is as a decorator in calvincontrol.

A test case for access of control interface is added.

Additionally, the user database now does not have to be configured via
the REST API. When the runtime now starts, it checks if the users.json file
contains passwords in clear, if it does, they are hashed and updated. This
allows us to bootstrap a users database with passwords in clear. Previously,
a users database could only be bootstrapped correclty via the REST API.
But when introducing access control on that REST API using the
same users database for authentication, things get awkward.

Signed-off-by: Hakan Englund <hakan.englund@ericsson.com>